requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-APScheduler==1.13.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
ollama==0.4.7
playwright==1.50.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
Werkzeug==3.1.3